added doDamage and takeDamage

+++ RPG_Project/src/main/java/org/example/items/Armor.java
@@ -25,7 +25,7 @@ public class Armor extends Item {
 
     @Override
     public int hashCode() {
-        return Objects.hashCode(protection);
+        return Objects.hashCode((Float) protection);
     }
 
     public float getProtection() {
